+++ Libplanet.Tests/Blockchain/NullPolicy.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Security.Cryptography;
 using Libplanet.Action;
@@ -22,6 +21,8 @@ namespace Libplanet.Tests.Blockchain
             _difficulty = difficulty;
         }
 
+        public ISet<Address> BlockedMiners { get; } = new HashSet<Address>();
+
         public IComparer<IBlockExcerpt> CanonicalChainComparer =>
             new TotalDifficultyComparer();
 
@@ -38,7 +39,20 @@ namespace Libplanet.Tests.Blockchain
             BlockChain<T> blockChain, Transaction<T> transaction) => null;
 
         public BlockPolicyViolationException ValidateNextBlock(
-            BlockChain<T> blocks, Block<T> nextBlock) => _exceptionToThrow;
+            BlockChain<T> blocks,
+            Block<T> nextBlock
+        )
+        {
+            if (_exceptionToThrow != null)
+            {
+                return _exceptionToThrow;
+            }
+
+            return BlockedMiners.Contains(nextBlock.Miner)
+                ? new BlockPolicyViolationException(
+                    $"Disallowed #{nextBlock.Index} {nextBlock.Hash} mined by {nextBlock.Miner}.")
+                : null;
+        }
 
         public int GetMaxBlockBytes(long index) => 1024 * 1024;